third_party/blink/web_tests/platform/generic/fast/events/constructors/webgl-context-event-constructor-expected.txt

This tests the constructor for the WebGLContextEvent DOM class.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS new WebGLContextEvent('eventType').bubbles is false
PASS new WebGLContextEvent('eventType').cancelable is false
PASS new WebGLContextEvent('eventType').statusMessage is ""
PASS new WebGLContextEvent('eventType', { bubbles: false }).bubbles is false
PASS new WebGLContextEvent('eventType', { bubbles: true }).bubbles is true
PASS new WebGLContextEvent('eventType', { cancelable: false }).cancelable is false
PASS new WebGLContextEvent('eventType', { cancelable: true }).cancelable is true
PASS new WebGLContextEvent('eventType', { statusMessage: 'abcde' }).statusMessage is "abcde"
PASS new WebGLContextEvent('eventType', { statusMessage: '' }).statusMessage is ""
PASS new WebGLContextEvent('eventType', { statusMessage: undefined }).statusMessage is ""
PASS new WebGLContextEvent('eventType', { statusMessage: null }).statusMessage is "null"
PASS new WebGLContextEvent('eventType', { statusMessage: false }).statusMessage is "false"
PASS new WebGLContextEvent('eventType', { statusMessage: true }).statusMessage is "true"
PASS new WebGLContextEvent('eventType', { statusMessage: 12345 }).statusMessage is "12345"
PASS new WebGLContextEvent('eventType', { statusMessage: 18446744073709551615 }).statusMessage is "18446744073709552000"
PASS new WebGLContextEvent('eventType', { statusMessage: NaN }).statusMessage is "NaN"
PASS new WebGLContextEvent('eventType', { statusMessage: [] }).statusMessage is ""
PASS new WebGLContextEvent('eventType', { statusMessage: [1, 2, 3] }).statusMessage is "1,2,3"
PASS new WebGLContextEvent('eventType', { statusMessage: {abcde: 12345} }).statusMessage is "[object Object]"
PASS new WebGLContextEvent('eventType', { statusMessage: {valueOf: function () { return 'abcde'; } } }).statusMessage is "[object Object]"
PASS new WebGLContextEvent('eventType', { bubbles: true, cancelable: true, statusMessage: 'abcde' }).bubbles is true
PASS new WebGLContextEvent('eventType', { bubbles: true, cancelable: true, statusMessage: 'abcde' }).cancelable is true
PASS new WebGLContextEvent('eventType', { bubbles: true, cancelable: true, statusMessage: 'abcde' }).statusMessage is 'abcde'
PASS successfullyParsed is true

TEST COMPLETE
